chore(tooling): rename route scripts to TypeScript
