+++ scripts/route-matrix.ts
